NEW - bug 308825: [Sequence Diagram] Time and Duration Support https://bugs.eclipse.org/bugs/show_bug.cgi?id=308825
Correct Orphan view Policy to work correctly with deletion of messages with linked time constraints in Sequence Diagram :
In a post commit context, the semantic hint of an orphan view is sometime "" (for example, create an async message, then delete it with ctrl-z).
In such a case, "" is not an integer and trying to create a new integer from it would throw an exception.