ASSIGNED - bug 315208: [Core] ResourceLoading stuff should be in its own pluggable/unpluggable project https://bugs.eclipse.org/bugs/show_bug.cgi?id=315208
The ResourceLoading has been moved in its own package.
A dedicated ModelSet exist, and is loaded as a service with a higher priority than the default service (key: oep.core.resource.ModelSet ).
The preferences have not been corrected. Part of preferences should be moved in the oep.core.resourceLoading project.