ASSIGNED - bug 308825: [Sequence Diagram] Time and Duration Support https://bugs.eclipse.org/bugs/show_bug.cgi?id=308825
Correct a bug for synchronizing moves of Time / Duration Constraints and Observations :
when a time element is put on the receive event of a syn message, take the start event instead of the message occurrence if time element is on the lifeline.

To achieve this, we must retain all possible occurrences atthe same location and chose the best one only at the last moment.