ASSIGNED - bug 274808: [Sequence Diagram] Improve Combined Fragment functionality https://bugs.eclipse.org/bugs/show_bug.cgi?id=274808
Fix - 2/ When a Combined Fragment is added in an Interaction Operand of a CF, the
new CF can not be distinguished from the Interaction Operand (and label is not
common to the two CF). Hence, when trying to add new Interaction Operands to
the new CF, they are added to the oldest one.