Fixing Bug 292124 - [General] Clean unused messages. The messages.properties have been cleaned for all UML diagrams.
Fixing a small compilation bug on state machine diagram editor. It was using the 1.6 method: String#isEmpty(), which does not exists in java 1.5

Fixing an exception thrown by the state machine diagram editor on startup, with a action bar contributor ID which was not found. The ID has been replaced by the common action bar contributor ID.