NEW - bug 321826: [Sequence Diagram] DurationConstraint to a message end is drawn on message https://bugs.eclipse.org/bugs/show_bug.cgi?id=321826
The problem was due to the fact that the creation was called on the message part instead of on the lifeline part.
When a duration element creation targets a message, the request is redirected by the tool to the nearest lifeline part.
In case we want to create a duration constraint on the message, the redirection which was performed just after still works.