NEW - bug 326421: [StateMachineDiagram] Composite and submachine States support https://bugs.eclipse.org/bugs/show_bug.cgi?id=326421
Corrects incorrect resize for state machines and composite states in general: it seems the constraint rectangle that is being passed on to the ChangeBoundsRequest is incorrect, which explains some of the gap shown on regions when resizing. Probably this rectangle is computed based on the figures and not the nodes, which is not what we need for state machine diagram. This should now be corrected.

Corrects drag and drop of existing regions on diagrams (when one re-arranges regions): subnodes are preserved.