NEW - bug 326554: [Sequence Diagram] DnD of message produces and error https://bugs.eclipse.org/bugs/show_bug.cgi?id=326554 ASSIGNED - bug 326775: [Sequence Diagram] General Ordering support https://bugs.eclipse.org/bugs/show_bug.cgi?id=326775
Handle Dnd of Messages and General Orderings.

Occurrence specifications are located at the correct location when the link is dropped.
This location is based on the order of interaction fragments, in order to preserve the same order as before.

Utility methods are also provided to work with fragments' order.

By the way, OccurenceSpecificationMoveHelper class has been renamed to OccurrenceSpecificationMoveHelper (missing 'r')