NEW - bug 325858: [Wizard template] Incorrect creation of model template when stereotyped elements are present https://bugs.eclipse.org/bugs/show_bug.cgi?id=325858
Bug is fixed with the use of 
		eObjectsToAdd = (List<EObject>)EcoreUtil.copyAll(templateResource.getContents());
in InitializeTemplateCommand instead of
		for(EObject eObject : templateResource.getContents()) {
			eObjectsToAdd.add(EcoreUtil.copy(eObject));
		}

What happens is that for profiled UML models, the resource contents are made of several eObjects, e.g. a Model and extra eObjects representing stereotypes, for instance SystemModel in an EASTADL model. The code used previously would deal with each eObject in turn, not building a complete HashMap of the keys/values for the whole file, thus references contained in the stereotypes were simply copied from the original file (default behavior when references are not found during the copy). The use of copyAll enables to trea the whole set of eObjects as a whole and keep track of all references. Which does the trick.