NEW - bug 330469: [Statemachine Diagram] Papyrus state machine diagram editor shall support Drag and Drop for transitions https://bugs.eclipse.org/bugs/show_bug.cgi?id=330469
A first support for transition drag and drop. It also adds support for drag and drop of composite states.

Limitations are the following:
- when one of the ends of a transition is missing on the diagram and should be create along with the transition, we forbid these ends to be in different regions or be composite states (or inner vertex of composite states)

Tod eal with these case you need to first create the ends, e.g. the composite states, etc. before dropping the transition.

Aside from this note that if a transition is a loop on the same state, thre drag and drop will wrongly create twice the same state, as is the case for the moment in the Activity diagram.

More advanced support is being investigated.