NEW - bug 330469: [Statemachine Diagram] Papyrus state machine diagram editor shall support Drag and Drop for transitions https://bugs.eclipse.org/bugs/show_bug.cgi?id=330469
Additional support for drag and drop.

To only limitation now is that the owner of the source/target of the link is present on the diagram, e.g. if the region owning a state which is the target of the link is not shown on the diagram, the drag and drop will not be allowed.

Also the duplication of state in case of looping transition is removed. This fix is also done for associations, etc. in the diagram.common plugin.

Besides, the patched code corrects some graphical malfunctions of the previous code, when showing/hiding information such as qualified names, changing fonts, etc. on state machines and composite states.