REOPENED - bug 307839: [Model explorer] Cannot move diagrams to different packages https://bugs.eclipse.org/bugs/show_bug.cgi?id=307839
Report modification from branch 0.8.X :
Fix by implementing diagram move. We check that the diagram can be created in the new location, with the correct type, and also that it is not a behavioral diagram before moving it.
Hence, the core interface ICreationCommand has been updated, but so its default implementations.