NEW - bug 357250: [General] Application reacts different ways when trying to modify a read-only resource https://bugs.eclipse.org/bugs/show_bug.cgi?id=357250
use a common command stack to execute command. This provides global undo context between views/editor.
use a CheckedOperationHistory instead of the DefaultOperationHistory : we can't specify a priority for the approvers in the default implementation, which is needed to handle SVN lock before file system read only attribute.

by me and Vincent Hemery