NEW - bug 372724: [Model Explorer] Search function does not traverse EReferences, LinkItems https://bugs.eclipse.org/bugs/show_bug.cgi?id=372724
The search function as currently implemented does not traverse EReferences and
LinkItems. This means that some tree items (for instance, a Class'
ownedAttributes) cannot be found.  

The comitted solution involves:
*A change to the NavigatorSearchDialog pattern search algorithm, allowing
traversal of problematic elements: this is done using an Adapter mechanism so
additional components contributing to the Model Explorer can easily extend it.
*A change to CustomCommonViewer and ModelExplorerView's reveal methods'
matching algorithm, with a complete MatchingItem refactoring. This overhaul
includes the creation of an IReferencable interface exposing a
getElementBehind() method, which additional components can implement to declare
their own match logic