Bug 396945 [Alf] Migration from Indigo to Juno: Regression in InstanceCreationExpression
In Indigo we had:
InstanceCreationExpression :
	'new' constructor=QualifiedNameWithBinding (tuple = Tuple | sequenceConstuctionCompletion = SequenceConstructionCompletion) (suffix = SuffixExpression) ?
;

In the migration to Juno, we had:
InstanceCreationExpression :
	'new' constructor=QualifiedNameWithBinding tuple = InstanceCreationTuple (suffix = SuffixExpression) ?
;

This commit includes corrections to have:
InstanceCreationExpression :
	'new' constructor=QualifiedNameWithBinding (tuple = InstanceCreationTuple (suffix = SuffixExpression) ? | sequenceConstuctionCompletion = SequenceConstructionCompletion)
;

Note that, as compared to Indigo version, this rule does not allow for suffixes in case of SequenceConstructionCompletion, which is more accurate regarding alf specification.
Note also that the StereotypeProperty editor has been regenerated to account for changes in the Alf grammar.