415371: Fix problems with equals() and hashCode() https://bugs.eclipse.org/bugs/show_bug.cgi?id=415371
Updated the equals() method to avoid resolving the source object, especially if it is an EObject, because it may no longer be available (e.g., after a CDO repository connection is closed).  There is still a problem because equality with an EObject is not symmetric, but I did not remove that because I don't know whether it is needed.

Also added the corresponding hashCode() method that was missing.