Bug 315231 - Improvement of xtext integration.
Applied patch from ITEMIS. Registers an EContentAdapter that observes
changes to the model outside of the xtext editor.
If the model changes, the annotation that stores the invalid string is
deleted.

There is one problem with this solution, the adapter is attached when
the DirectEditManager opens the first time. 
This should be done when the ResourceSet or the Editing domain is
created.