415373: [CDO] Add an Object Localizer Service https://bugs.eclipse.org/bugs/show_bug.cgi?id=415373
The CDO integration layer needs to be able to override the mechanism by which the diagram's drop-target listener finds the correspondents of objects from other resource sets in its own resource set (let's call this "object localization").  The drop-target listener now delegates this localization function to a new service, which can be overridden by CDO integration.