415373: [CDO] Ensure proper cleanup of repositories in tests https://bugs.eclipse.org/bugs/show_bug.cgi?id=415373
Fix test failures on "repository already exists" condition caused by the repository registry being saved when a repository is connected, to remember its UUID for the next Eclipse session.

Also ensure that repositories have access (through the managed container) to the factory that created them, for this saving of the registry and other purposes.