Bug 315231 - Improvement of xtext integration.
Applied patches from ITEMIS. Disables the direct editing for mouse
actions and enables it for Keyboard actions. Unfortunately, it was not
possible to set the key to 'F2' because the 'rename element' dialog
consumes the key pressed event in some way that the tree editor does not
get it. The key is set to 'F3' now, the commiter who implemented the
rename element dialog should have a look if he can disable this behavior
for elements with an xtext grammar. Then we can simply change the SWT.F3
to SWT.F2.