402525: [Widgets / Transactions] Papyrus dialogs should be transactional https://bugs.eclipse.org/bugs/show_bug.cgi?id=402525
Introduce an IAtomicOperationExecutor adapter for the object being edited by the widgets to to execute complex operations on the EMF command-stack (for EObjects being edited).  This command stack is understood to support nested command execution.

Add the element being edited as context for the ReferenceValueFactory's creation of a new referenced object.  This provides resource-set context for nested editors that need to create further new elements wherever in the model.  This resource-set also can be used to look up the LabelProviderService for presentation of objects in the dialogs.