Apply bugfox from B. Maggi on master (already applied on branch (see commit id in following text)
e8fb604a0762987da518649578bbee6308742323 Use DiagramCommandStack to
allow editPart creation notification
(used when dropping a link create a view for a (source or target) node)



58f0eb6713f732753ab262d88cd13b2464aeadcc Bug 427271 - [SysML BDD] Can no
more draw associations between a Block
and an Actor. 

f86c1d8b36eb37cee0122e4de920ed71a6f952bc Bug 406290 - [SysML IBD] When
the type of a part or a reference changes

the nestedConnectors linked to an internal component of the
part/refrence are hidden but are not deleted from the model.

When destroying a part or changing its type, destroying all connectors
that reference this part in NestedConnectorEnd.