323802: [General] Papyrus shall provide a read only mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=323802
Implement a custom TransactionChangeRecorder that aborting the transaction if a read-only object is modified.

Add a new extension point oep.infra.gmfdiag.commands.historyListeners for registration of listeners to be added to the Papyrus operation history.  Implement a listener that detects operations that failed because of rollback and presents a status dialog explaining the rollback.

Add a new label provider for EMF Resources to support presentation of the rollback dialog when an affected object is a resource.