323802: [General] Papyrus shall provide a read only mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=323802
Merely disapproving the edit in an advice is not enough because in the diagram editor, the GEF EditPolicy can produce a command that is used instead of the "semantic" command produced (and null if disapproved) by the edit-helper.  So, we must actually return an unelectable command in order to block the execution of this fall-back command.