323802: [General] Papyrus shall provide a read only mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=323802
Support editing of read-only resources (with the approver's prompt to make resources writable) in the Model Explorer and other contexts than diagrams.

This involves being more aggressive in trying to extract the "affected files" from IUndoableOperations in the approver, because drag-and-drop operations (for example) in the Model Explorer are not GMF commands, but rather include GMF commands nested within wrappers and composites.