323802: [General] Papyrus shall provide a read only mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=323802
Make the contents of referenced model resources read-only by default, via a new read-only handler.  A referenced model resource is one that is not a constituent of the "main" model (being edited in the current ModelSet).  This includes controlled sub-units.  To determine efficiently whether a resource is a sub-unit of the main model, we add an adapter that tracks the topology of the sub-unit structure by detecting cross-resource-containment changes and caching the child-unit/parent-unit relationships.

The new read-only handler makes resources writable (with user prompt) by remembering their URIs locally.  This memory is not retained if the model editor is closed and then opened again.