323802: [General] Papyrus shall provide a read only mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=323802
Update the Papyrus editing domain to prompt to make resources writable before rolling back (and then only if cancelled or no handler can make them writable).

Also make the ReadOnlyManager properly short-circuit subsequent handlers if a handler that is requested to make resources writable does not (additional prompts to the user are useless because the result is already determined false).