429242: [CDO] [Resource Management] Refactoring of the 3-files model and PageManager https://bugs.eclipse.org/bugs/show_bug.cgi?id=429242
The isLegacy(URI) method did not account for the fact that some models have two file extensions:  my.profile.di.  Trimming the file extension again is not idempotent in these cases and resulted in a false negative result.  Consequently, the *.sash resource would end up being created in the same store (workspace/CDO) as the model instead of in the workspace metadata area.