429826: [Read-Only] Orthogonal Classification of Read-Only Concerns https://bugs.eclipse.org/bugs/show_bug.cgi?id=429826
Update clients of the ReadOnlyManager to ensure that all read-only axes are considered where appropriate (such as in the EnableWriteCommandHandler).

Add a listener protocol to notify interested parties when the read-only state of a resource changes on some axis.  Update the Model Explorer view to trigger refresh on read-only state changes.