429826: [Read-Only] Orthogonal Classification of Read-Only Concerns https://bugs.eclipse.org/bugs/show_bug.cgi?id=429826
The read-only handler for sash models short-circuits the read-only test for the composite Papyrus model resource because it indicates that definitively none of the resources are read-only when another lower-priority handler would answer that some of them are.

This is fixed by answering a negative only when it applies to all resources in the query.