430726: [Class Diagram] Drag-n-drop from Model Explorer to Diagram is broken post M6 https://bugs.eclipse.org/bugs/show_bug.cgi?id=430726
Update the diagram editor template to access the selection directly from the LocalSelectionTransfer, because otherwise the transfer data is null and the log is spammed with "Received wrong transfer data" messages.

The Class Diagram plug-in is regenerated.  Because there would otherwise be more than 580 files changed only in formatting differences, only the UMLDiagramEditor class is committed.

The xtend/ folder encoding is set to ISO-8859-1 to ensure that the Xpand guillemets are properly decoded.