430880: [Generic Table] Cannot drag-and-drop into generic table on Mac https://bugs.eclipse.org/bugs/show_bug.cgi?id=430880
Handle LocalSelectionTransfer as well as LocalTransfer (which latter does not contain data on Mac until drop time).

Also, handle the new EMF Facet 0.2-based tree-wrappers in the Model Explorer that are no longer adaptables.

Finally, set the drop-target event detail to DROP_COPY as in diagrams when the drop is enabled, to ensure that the user feedback indicates (in the platform's appropriate fashion) that copy semantics are enforced (nothing will be removed from the drag source).