Bug 425208: Duplicate visibility keywords
A Java utility function is used to merge adjacent visibitily sections
(within a C++ class definition).  For example, this:

    class T
    {
    public: void func1();
    public: void func2();
    };

Should appear as:

    class T
    {
    public: void func1();
            void func2();
    };

This Java utility maintains internal state so that the visibility
keyword is written only once.

This problem was raised because the internal state is updated even when
the keyword is not written to the file.  For example, this happens if
the generation templates try to generate an empty section with a
different visibility from what was currently written.

This patch modifies the utility function so that internal state is only
updated when the keyword is actually written out to the file.

Change-Id: I052ede33110d382409476cac4cef38a417aaa904
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>