431953: Stereotype garbage left in .uml file after removing profile (crash reason?) https://bugs.eclipse.org/bugs/show_bug.cgi?id=431953
Refactor the initialization of the ModelSet service so that its own service factory associates it with the service registry (removing the additional service-factory that did this).

This introduces a new concept of service adapters that can externally provide a POJO service instance with the service lifecycle hooks, for POJOs that are injected into the service registry by a client and not created by a factory.  Non-POJO services and factory-created services do not need this because they have the service lifecycle built-in.

Update the CoreMultiDiagramEditor to start itself early to let other services such as the model set find it in the registry.  Fix an NPE in the abstract editor triggered by the service registry's attempt to obtain a service adapter from it.