431953: Stereotype garbage left in .uml file after removing profile (crash reason?) https://bugs.eclipse.org/bugs/show_bug.cgi?id=431953
Implement a model-set snippet that detects and repairs (with user interaction) stereotype applications that either are not described by any profile application in the context of the a resource's root element when a resource is loaded.

Includes refactoring of the UI and back-end of the Switch Profiles functionality.

Cases of broken stereotypes include:
- instances of EClasses from a different version of the profile's Ecore definition than what is currently applied
- instances of EClasses from a profile that is not applied at all
- instances of EClasses from a profile that may be applied but which is unresolved (no longer exists at its former location)